Add SwatchCard component and refactor ContrastChecker to utilize it
